Created add-in for macro
Created an add-in file with a custom ribbon and button (Icon MSO:TableSave). The button executes the `ExportRangeToDataFrame` macro for the current selection.

Added Save As dialog box for user to specify destination and file name for the dataframe pickle file.

VBA functions now dynamically get the export path and active spreadsheet path and include them in the python script execution command string.

Python script now takes export path and active spreadsheet path as arguments and uses them for loading and serializing the data.